requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LLM
google-genai>=0.3.0
tenacity==8.2.3

# RAG & Embeddings
sentence-transformers==2.3.1
numpy==1.26.3

# Phase 2 (optional)
# chromadb==0.4.22
# sentence-transformers[onnx]==2.3.1

# Database
aiosqlite==0.19.0

# Utilities
tiktoken==0.5.2
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# Development
black==24.1.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
